// src/app/modules/dashboard/referral/models/request/request-update-management-referral.model.ts
import { PhoneCodeEnum } from 'src/app/shared/enums';
import { ManagementReferralReasonEnum } from '../../enums';

export interface RequestUpdateManagementReferralModel {
  id: string;
  employeeName: string;
  dOB: string;
  department: string;
  employeeNumberCode: PhoneCodeEnum;
  employeeNumber: string;
  jobTitle: string;
  workLocation: string;
  serviceLength: string;
  contactEmail: string;
  homeAddress: string;
  contactTelephoneNumberCode: PhoneCodeEnum;
  contactTelephoneNumber: string;
  reason: ManagementReferralReasonEnum;
  isCurrentlyAbsent: boolean;
  absenceStartDate: string;
  absenceReason?: string;
  isCurrentlyFitForWork: boolean;
  areThereAttendaceRecordReasons: boolean;
  areTherePerformanceConsernsReasons: boolean;
  haveWorkTasksContributed: boolean;
  haveDomesticIssuesContributed: boolean;
  whatAreRecoveryTimescales: boolean;
  areWorkRestrictionsRequired: boolean;
  willRestrictionsBeTemporary: boolean;
  isIndividualTakingSteps: boolean;
  whatFactorsMightDelay: boolean;
  doesIndividualMeetCriteria: boolean;
  isMedConditionWillBeCovered: boolean;
  whatIsLongtermOutlook: boolean;
  extraPoints: string;
  usingDisplayScreen: boolean;
  safetyCriticalJob: boolean;
  officeBased: boolean;
  workingAtHeights: boolean;
  regularPublicContact: boolean;
  confinedSpaces: boolean;
  nightWorker: boolean;
  loneWorker: boolean;
  vibrationTools: boolean;
  driving: boolean;
  noisyEnvironments: boolean;
  operatingMachinery: boolean;
  dustChems: boolean;
  manualHandling: boolean;
  workDuties: string;
  documentsWithReferral: string;
  isReferralReasonWasDiscussed: boolean;
  referringManagerName: string;
  contactTelephone: string;
  date: string;
  isSubmit: boolean;
  productId: string;
}